% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram for a Numeric Column in a Data Frame}
\usage{
plot_histogram(data, column, title)
}
\arguments{
\item{data}{A data frame.}

\item{column}{A string. The name of the numeric column to plot.}
}
\value{
A ggplot2 histogram object.
}
\description{
This function takes a data frame and a column name and plots a histogram of the selected column.
}
\examples{
plot_histogram(mtcars, "mpg")

}
